refactor: 레이아웃 공통 컴포넌트 분리
