+++ frontend/src/components/Main/GridViewCoupons.tsx
