+++ frontend/src/components/shared/Body.tsx
@@ -1,0 +1,1 @@
+import styled from '@emotion/styled';
