+++ frontend/src/components/shared/Header.tsx
@@ -1,0 +1,12 @@
+import styled from '@emotion/styled';
+
+const Header = styled.header`
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 15px;
+  color: white;
+  margin: 10px 0 0 2vw;
+`;
+
+export default Header;
