+++ frontend/src/components/shared/HeaderText.tsx
@@ -1,0 +1,8 @@
+import styled from '@emotion/styled';
+
+const HeaderText = styled.p`
+  font-size: 22px;
+  margin-left: calc(1vw + 6px);
+`;
+
+export default HeaderText;
