+++ frontend/src/components/shared/PageLayout.tsx
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled';
+
+const PageLayout = styled.section`
+  display: flex;
+  flex-direction: column;
+  padding: 5px;
+`;
+
+export default PageLayout;
