+++ frontend/src/pages/SelectReceiver.tsx
@@ -1,25 +1,29 @@
-import ArrowBackButton from '../commons/Header/ArrowBackButton';
+import ArrowBackButton from '../components/shared/ArrowBackButton';
 import styled from '@emotion/styled';
 import useSelectReceiver from '../hooks/SelectReceiver/useSelectReceiver';
-import UserSearchInput from '../commons/SelectReceiver/UserSearchInput';
-import ListViewUsers from '../commons/SelectReceiver/ListViewUsers';
-import CheckedUsers from '../commons/SelectReceiver/CheckedUsers';
+import UserSearchInput from '../components/SelectReceiver/UserSearchInput';
+import ListViewUsers from '../components/SelectReceiver/ListViewUsers';
+import CheckedUsers from '../components/SelectReceiver/CheckedUsers';
 import ArrowForwardIosIcon from '@mui/icons-material/ArrowForwardIos';
 import { Link } from 'react-router-dom';
 import { css } from '@emotion/react';
+
+import PageLayout from '../components/shared/PageLayout';
+import Header from '../components/shared/Header';
+import HeaderText from '../components/shared/HeaderText';
 
 const SelectReceiver = () => {
   const { users, isLoading, error, checkedUsers, toggleUser, uncheckUser, isCheckedUser } =
     useSelectReceiver();
 
   return (
-    <S.Container>
-      <S.Header>
+    <PageLayout>
+      <Header>
         <Link to='/'>
           <ArrowBackButton />
         </Link>
-        <S.HeaderText>누구한테 보낼까요?</S.HeaderText>
-      </S.Header>
+        <HeaderText>누구한테 보낼까요?</HeaderText>
+      </Header>
       <S.Body>
         {checkedUsers.length !== 0 && (
           <CheckedUsers checkedUsers={checkedUsers} onClickDelete={uncheckUser} />
@@ -39,7 +43,7 @@ const SelectReceiver = () => {
           <ArrowForwardIosIcon />
         </S.LongButton>
       </S.SendButtonBox>
-    </S.Container>
+    </PageLayout>
   );
 };
 
@@ -48,23 +52,6 @@ type ButtonProps = {
 };
 
 const S = {
-  Container: styled.div`
-    display: flex;
-    flex-direction: column;
-    padding: 5px;
-  `,
-  Header: styled.header`
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 15px;
-    color: white;
-    margin: 10px 0 0 2vw;
-  `,
-  HeaderText: styled.p`
-    font-size: 26px;
-    margin-left: calc(1vw + 6px);
-  `,
   Body: styled.div`
     display: flex;
     flex-direction: column;
